Add two necessary hooks.
One hook applies a context when trying to retrieve a pointer from an
array when accessing a `Memory` object.

This is necessary for CachedArrays in the case where the memory encloses
a `NotBusy` CachedArray. By providing this context, we can perform the
appropriate actions at the CachedArrays level.

The second hook we add is a kernel exit hook which is called on the
return of a OneDNN kernel invocation.

For CachedArrays, this allows us to release the returned encapsulated
CachedArray.

TODO: Right now, changing the type of the CachedArray inside a memory is
pretty awkward. It would be a bit easier if the actual `Lib.dnnl_memory_t`
pointer was an independantly managed object that could be referenced
multiple times, much like a CachedArrays `Region`.